Refactor forward model names. Code cleaning and optimizations. API in Game.
